extract edge/serverless handler logic into separate files
