+++ packages/node/src/edge-functions/edge-wasm-plugin.ts
