+++ packages/node/src/utils.ts
@@ -1,5 +1,6 @@
 import { extname } from 'path';
 import { pathToRegexp } from 'path-to-regexp';
+import { debug } from '@vercel/build-utils';
 
 export function getRegExpFromMatchers(matcherOrMatchers: unknown): string {
   if (!matcherOrMatchers) {
@@ -54,4 +55,15 @@ export function entrypointToOutputPath(
     return entrypoint.slice(0, entrypoint.length - ext.length);
   }
   return entrypoint;
+}
+
+export function logError(error: Error) {
+  console.error(error.message);
+  if (error.stack) {
+    // only show the stack trace if debug is enabled
+    // because it points to internals, not user code
+    const errorPrefixLength = 'Error: '.length;
+    const errorMessageLength = errorPrefixLength + error.message.length;
+    debug(error.stack.substring(errorMessageLength + 1));
+  }
 }
